fix: execution engine child task ordering, file size violations, docs (#1223 phases 3-4)
