+++ src/synthorg/engine/workflow/execution_lifecycle.py
@@ -1,0 +1,386 @@
+"""Workflow execution lifecycle transitions and task-event handling.
+
+Extracted from ``execution_service.py`` to keep file sizes manageable.
+All functions operate on ``WorkflowExecution`` models and delegate
+persistence to the injected repository.
+"""
+
+from datetime import UTC, datetime
+
+from synthorg.core.enums import (
+    TaskStatus,
+    WorkflowExecutionStatus,
+    WorkflowNodeExecutionStatus,
+    WorkflowNodeType,
+)
+from synthorg.engine.errors import (
+    WorkflowExecutionError,
+    WorkflowExecutionNotFoundError,
+)
+from synthorg.engine.task_engine_models import TaskStateChanged  # noqa: TC001
+from synthorg.engine.workflow.execution_models import (  # noqa: TC001
+    WorkflowExecution,
+    WorkflowNodeExecution,
+)
+from synthorg.observability import get_logger
+from synthorg.observability.events.workflow_execution import (
+    WORKFLOW_EXEC_CANCELLED,
+    WORKFLOW_EXEC_COMPLETED,
+    WORKFLOW_EXEC_FAILED,
+    WORKFLOW_EXEC_INVALID_STATUS,
+    WORKFLOW_EXEC_NODE_TASK_COMPLETED,
+    WORKFLOW_EXEC_NODE_TASK_FAILED,
+    WORKFLOW_EXEC_NOT_FOUND,
+)
+from synthorg.persistence.errors import VersionConflictError
+from synthorg.persistence.workflow_execution_repo import (  # noqa: TC001
+    WorkflowExecutionRepository,
+)
+
+logger = get_logger(__name__)
+
+_TERMINAL_TASK_STATUSES = frozenset(
+    {TaskStatus.COMPLETED, TaskStatus.FAILED, TaskStatus.CANCELLED},
+)
+
+
+# -- CRUD helpers ----------------------------------------------------------
+
+
+async def get_execution(
+    repo: WorkflowExecutionRepository,
+    execution_id: str,
+) -> WorkflowExecution | None:
+    """Retrieve a workflow execution by ID."""
+    return await repo.get(execution_id)
+
+
+async def list_executions(
+    repo: WorkflowExecutionRepository,
+    definition_id: str,
+) -> tuple[WorkflowExecution, ...]:
+    """List executions for a workflow definition."""
+    return await repo.list_by_definition(definition_id)
+
+
+async def cancel_execution(
+    repo: WorkflowExecutionRepository,
+    execution_id: str,
+    *,
+    cancelled_by: str,
+) -> WorkflowExecution:
+    """Cancel a workflow execution.
+
+    Raises:
+        WorkflowExecutionNotFoundError: If not found.
+        WorkflowExecutionError: If execution is already terminal.
+    """
+    execution = await repo.get(execution_id)
+    if execution is None:
+        logger.warning(
+            WORKFLOW_EXEC_NOT_FOUND,
+            execution_id=execution_id,
+        )
+        msg = f"Workflow execution {execution_id!r} not found"
+        raise WorkflowExecutionNotFoundError(msg)
+
+    terminal_statuses = {
+        WorkflowExecutionStatus.COMPLETED,
+        WorkflowExecutionStatus.FAILED,
+        WorkflowExecutionStatus.CANCELLED,
+    }
+    if execution.status in terminal_statuses:
+        msg = (
+            f"Cannot cancel execution {execution_id!r}"
+            f" in terminal status {execution.status.value!r}"
+        )
+        logger.warning(
+            WORKFLOW_EXEC_CANCELLED,
+            execution_id=execution_id,
+            error=msg,
+        )
+        raise WorkflowExecutionError(msg)
+
+    now = datetime.now(UTC)
+    cancelled = execution.model_copy(
+        update={
+            "status": WorkflowExecutionStatus.CANCELLED,
+            "updated_at": now,
+            "completed_at": now,
+            "version": execution.version + 1,
+        }
+    )
+    await repo.save(cancelled)
+
+    logger.info(
+        WORKFLOW_EXEC_CANCELLED,
+        execution_id=execution_id,
+        cancelled_by=cancelled_by,
+    )
+
+    return cancelled
+
+
+async def complete_execution(
+    repo: WorkflowExecutionRepository,
+    execution_id: str,
+) -> WorkflowExecution:
+    """Transition a running execution to COMPLETED.
+
+    Raises:
+        WorkflowExecutionNotFoundError: If not found.
+        WorkflowExecutionError: If execution is not RUNNING.
+    """
+    execution = await _load_running(repo, execution_id)
+    now = datetime.now(UTC)
+    completed = execution.model_copy(
+        update={
+            "status": WorkflowExecutionStatus.COMPLETED,
+            "updated_at": now,
+            "completed_at": now,
+            "version": execution.version + 1,
+        },
+    )
+    await repo.save(completed)
+    logger.info(
+        WORKFLOW_EXEC_COMPLETED,
+        execution_id=execution_id,
+    )
+    return completed
+
+
+async def fail_execution(
+    repo: WorkflowExecutionRepository,
+    execution_id: str,
+    *,
+    error: str,
+) -> WorkflowExecution:
+    """Transition a running execution to FAILED.
+
+    Raises:
+        WorkflowExecutionNotFoundError: If not found.
+        WorkflowExecutionError: If execution is not RUNNING.
+    """
+    execution = await _load_running(repo, execution_id)
+    now = datetime.now(UTC)
+    failed = execution.model_copy(
+        update={
+            "status": WorkflowExecutionStatus.FAILED,
+            "error": error,
+            "updated_at": now,
+            "completed_at": now,
+            "version": execution.version + 1,
+        },
+    )
+    await repo.save(failed)
+    logger.info(
+        WORKFLOW_EXEC_FAILED,
+        execution_id=execution_id,
+        error=error,
+    )
+    return failed
+
+
+# -- Task-event handling ---------------------------------------------------
+
+
+async def handle_task_state_changed(
+    repo: WorkflowExecutionRepository,
+    event: TaskStateChanged,
+) -> None:
+    """React to a task state change from the TaskEngine.
+
+    Correlates the task to a running workflow execution and
+    transitions the execution to COMPLETED or FAILED as
+    appropriate.
+    """
+    if event.mutation_type != "transition":
+        return
+    if event.new_status not in _TERMINAL_TASK_STATUSES:
+        return
+
+    execution = await repo.find_by_task_id(event.task_id)
+    if execution is None:
+        return
+
+    try:
+        if event.new_status in {TaskStatus.FAILED, TaskStatus.CANCELLED}:
+            await _handle_task_failed(repo, execution, event)
+        else:
+            await _handle_task_completed(repo, execution, event)
+    except VersionConflictError:
+        retry_event = (
+            WORKFLOW_EXEC_NODE_TASK_FAILED
+            if event.new_status in {TaskStatus.FAILED, TaskStatus.CANCELLED}
+            else WORKFLOW_EXEC_NODE_TASK_COMPLETED
+        )
+        logger.warning(
+            retry_event,
+            execution_id=execution.id,
+            task_id=event.task_id,
+            error="Concurrent modification; re-fetching execution",
+        )
+        refreshed = await repo.get(execution.id)
+        if refreshed is None:
+            return
+        if refreshed.status in {
+            WorkflowExecutionStatus.COMPLETED,
+            WorkflowExecutionStatus.FAILED,
+            WorkflowExecutionStatus.CANCELLED,
+        }:
+            return
+        if event.new_status in {TaskStatus.FAILED, TaskStatus.CANCELLED}:
+            await _handle_task_failed(repo, refreshed, event)
+        else:
+            await _handle_task_completed(repo, refreshed, event)
+
+
+# -- Private helpers -------------------------------------------------------
+
+
+async def _handle_task_failed(
+    repo: WorkflowExecutionRepository,
+    execution: WorkflowExecution,
+    event: TaskStateChanged,
+) -> None:
+    """Handle a task failure or cancellation event."""
+    updated = _update_node_status(
+        execution,
+        event.task_id,
+        WorkflowNodeExecutionStatus.TASK_FAILED,
+    )
+    now = datetime.now(UTC)
+    verb = "cancelled" if event.new_status is TaskStatus.CANCELLED else "failed"
+    error_msg = f"Task {event.task_id} {verb}"
+    failed = updated.model_copy(
+        update={
+            "status": WorkflowExecutionStatus.FAILED,
+            "error": error_msg,
+            "updated_at": now,
+            "completed_at": now,
+        },
+    )
+    await repo.save(failed)
+    logger.info(
+        WORKFLOW_EXEC_NODE_TASK_FAILED,
+        execution_id=execution.id,
+        task_id=event.task_id,
+    )
+    logger.info(
+        WORKFLOW_EXEC_FAILED,
+        execution_id=execution.id,
+        error=error_msg,
+    )
+
+
+async def _handle_task_completed(
+    repo: WorkflowExecutionRepository,
+    execution: WorkflowExecution,
+    event: TaskStateChanged,
+) -> None:
+    """Handle a task completion event."""
+    updated = _update_node_status(
+        execution,
+        event.task_id,
+        WorkflowNodeExecutionStatus.TASK_COMPLETED,
+    )
+    logger.info(
+        WORKFLOW_EXEC_NODE_TASK_COMPLETED,
+        execution_id=execution.id,
+        task_id=event.task_id,
+    )
+    if _all_tasks_completed(updated):
+        now = datetime.now(UTC)
+        completed = updated.model_copy(
+            update={
+                "status": WorkflowExecutionStatus.COMPLETED,
+                "updated_at": now,
+                "completed_at": now,
+            },
+        )
+        await repo.save(completed)
+        logger.info(
+            WORKFLOW_EXEC_COMPLETED,
+            execution_id=execution.id,
+        )
+    else:
+        await repo.save(updated)
+
+
+async def _load_running(
+    repo: WorkflowExecutionRepository,
+    execution_id: str,
+) -> WorkflowExecution:
+    """Load an execution and validate it is RUNNING."""
+    execution = await repo.get(execution_id)
+    if execution is None:
+        logger.warning(
+            WORKFLOW_EXEC_NOT_FOUND,
+            execution_id=execution_id,
+        )
+        msg = f"Workflow execution {execution_id!r} not found"
+        raise WorkflowExecutionNotFoundError(msg)
+
+    if execution.status is not WorkflowExecutionStatus.RUNNING:
+        msg = (
+            f"Cannot transition execution {execution_id!r}"
+            f" in status {execution.status.value!r}"
+            " (expected 'running')"
+        )
+        logger.warning(
+            WORKFLOW_EXEC_INVALID_STATUS,
+            execution_id=execution_id,
+            current_status=execution.status.value,
+            error=msg,
+        )
+        raise WorkflowExecutionError(msg)
+
+    return execution
+
+
+def _update_node_status(
+    execution: WorkflowExecution,
+    task_id: str,
+    new_status: WorkflowNodeExecutionStatus,
+) -> WorkflowExecution:
+    """Return a copy with one node's status updated."""
+    found = False
+    updated_nodes: list[WorkflowNodeExecution] = []
+    for ne in execution.node_executions:
+        if ne.task_id == task_id:
+            updated_nodes.append(
+                ne.model_copy(update={"status": new_status}),
+            )
+            found = True
+        else:
+            updated_nodes.append(ne)
+
+    if not found:
+        msg = f"task_id {task_id!r} not found in execution {execution.id!r}"
+        logger.warning(
+            WORKFLOW_EXEC_NOT_FOUND,
+            execution_id=execution.id,
+            task_id=task_id,
+            error=msg,
+        )
+        raise ValueError(msg)
+
+    return execution.model_copy(
+        update={
+            "node_executions": tuple(updated_nodes),
+            "updated_at": datetime.now(UTC),
+            "version": execution.version + 1,
+        },
+    )
+
+
+def _all_tasks_completed(execution: WorkflowExecution) -> bool:
+    """Check if all non-skipped TASK nodes have completed."""
+    for ne in execution.node_executions:
+        if ne.node_type is not WorkflowNodeType.TASK:
+            continue
+        if ne.status is WorkflowNodeExecutionStatus.SKIPPED:
+            continue
+        if ne.status is not WorkflowNodeExecutionStatus.TASK_COMPLETED:
+            return False
+    return True
